Move date helpers to utils/ and use in edit page
